+++ docs/CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
+- World generation module.
+- Map water gen with flowing rivers.
 - UI build menu module.
 - UI tooltips module.
 - UI roads module.